+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Initial release of eslint-plugin-better-cap-config
-- Rule: `no-deprecated-cds-fiori-draft-compat`
+- Rule: `cds-fiori-draft-compat`
 - Rule: `cds-features-odata-new-adapter`
 - Rule: `cds-features-cds-validate`
 - Support for ESLint v9+ flat configuration